[IIIF-497] :construction: WIP second pass rename - this will break other stuff
