+++ .travis.yml
@@ -14,7 +14,7 @@ jobs:
       install:
         - mvn install -q -Dmaven.test.skip -Dmaven.javadoc.skip -B -V
       script:
-        - mvn verify -Dmanifeststore.s3.access_key="${AWS_ACCESS_KEY_ID}" -Dmanifeststore.s3.secret_key="${AWS_SECRET_ACCESS_KEY}"
+        - mvn verify -Dfester.s3.access_key="${AWS_ACCESS_KEY_ID}" -Dfester.s3.secret_key="${AWS_SECRET_ACCESS_KEY}"
       # Cache the local Maven repo to save time
       cache:
         directories:
@@ -26,7 +26,7 @@ jobs:
       deploy:
         - provider: script
           script:
-            - MANIFESTSTORE_TAG=latest .travis/trigger.sh uclalibrary/docker-fester
+            - FESTER_TAG=latest .travis/trigger.sh uclalibrary/docker-fester
           on:
             branch: master
       if: (NOT type IN (pull_request)) AND (branch = master)
@@ -36,7 +36,7 @@ jobs:
       deploy:
         - provider: script
           script:
-            - MANIFESTSTORE_TAG=$TRAVIS_TAG .travis/trigger.sh uclalibrary/docker-fester
+            - FESTER_TAG=$TRAVIS_TAG .travis/trigger.sh uclalibrary/docker-fester
           on:
             tags: true
       if: (NOT type IN (pull_request)) AND (branch =~ /^(0|[1-9]\d*)\.(0|[1-9]\d*)\.(0|[1-9]\d*)(-(0|[1-9]\d*|\d*[a-zA-Z-][0-9a-zA-Z-]*)(\.(0|[1-9]\d*|\d*[a-zA-Z-][0-9a-zA-Z-]*))*)?(\+[0-9a-zA-Z-]+(\.[0-9a-zA-Z-]+)*)?$/)
