+++ README.md
@@ -4,12 +4,12 @@ A IIIF manifest storage microservice. It will provide almost-full CRUD access (c
 
 ## Configuring the Build
 
-The IIIF manifest store uses an S3 bucket for back-end storage. To be able to run the project's tests, several configuration values must be supplied:
+Fester uses an S3 bucket for back-end storage. To be able to run the project's tests, several configuration values must be supplied:
 
-* manifeststore.s3.bucket
-* manifeststore.s3.access_key
-* manifeststore.s3.secret_key
-* manifeststore.s3.region
+* fester.s3.bucket
+* fester.s3.access_key
+* fester.s3.secret_key
+* fester.s3.region
 
 These values can be set as properties in your system's Maven settings.xml file (or be supplied on the command line at build time).
