+++ pom.xml
@@ -61,16 +61,16 @@
     <jcip.annotations.version>1.0-1</jcip.annotations.version>
 
     <!-- Name of the main Vert.x verticle -->
-    <main.verticle>edu.ucla.library.iiif.manifeststore.verticles.MainVerticle</main.verticle>
+    <main.verticle>edu.ucla.library.iiif.fester.verticles.MainVerticle</main.verticle>
 
     <!-- Whether the test instances started with the 'live' profile reloads code automatically -->
     <live.test.reloads>false</live.test.reloads>
 
     <!-- AWS configuration properties -->
-    <manifeststore.s3.region>us-east-1</manifeststore.s3.region>
-    <manifeststore.s3.bucket>iiif-fester</manifeststore.s3.bucket>
-    <manifeststore.s3.access_key>YOUR_ACCESS_KEY</manifeststore.s3.access_key>
-    <manifeststore.s3.secret_key>YOUR_SECRET_KEY</manifeststore.s3.secret_key>
+    <fester.s3.region>us-east-1</fester.s3.region>
+    <fester.s3.bucket>iiif-fester</fester.s3.bucket>
+    <fester.s3.access_key>YOUR_ACCESS_KEY</fester.s3.access_key>
+    <fester.s3.secret_key>YOUR_SECRET_KEY</fester.s3.secret_key>
   </properties>
 
   <dependencies>
@@ -169,7 +169,7 @@
       </plugin>
       <!--<![CDATA[
         To run the generate-codes plugin outside of a full build, run:
-         mvn info.freelibrary:freelib-utils:generate-codes -DmessageFiles=src/main/resources/manifeststore_messages.xml
+         mvn info.freelibrary:freelib-utils:generate-codes -DmessageFiles=src/main/resources/fester_messages.xml
         Bash shell users can run the shorter form:
          mvn info.freelibrary:freelib-utils:generate-codes -DmessageFiles=$(find src -name *_messages.xml)
       ]]>-->
@@ -186,7 +186,7 @@
             <configuration>
               <generatedSourcesDirectory>${project.basedir}/src/main/generated</generatedSourcesDirectory>
               <messageFiles>
-                <messageFile>${project.basedir}/src/main/resources/manifeststore_messages.xml</messageFile>
+                <messageFile>${project.basedir}/src/main/resources/fester_messages.xml</messageFile>
               </messageFiles>
             </configuration>
           </execution>
@@ -211,7 +211,7 @@
                 <resource>
                   <directory>${basedir}/src/main/resources</directory>
                   <includes>
-                    <include>manifeststore.yaml</include>
+                    <include>fester.yaml</include>
                   </includes>
                 </resource>
               </resources>
