+++ src/main/java/edu/ucla/library/iiif/manifeststore/Config.java
@@ -1,5 +1,5 @@
 
-package edu.ucla.library.iiif.manifeststore;
+package edu.ucla.library.iiif.fester;
 
 /**
  * Some constant values for the project.
@@ -12,13 +12,13 @@ public final class Config {
     // The OpenAPI specification configuration parameter
     public static final String OPENAPI_SPEC_PATH = "openapi.spec.path";
 
-    public static final String S3_ACCESS_KEY = "manifeststore.s3.access_key";
+    public static final String S3_ACCESS_KEY = "fester.s3.access_key";
 
-    public static final String S3_SECRET_KEY = "manifeststore.s3.secret_key";
+    public static final String S3_SECRET_KEY = "fester.s3.secret_key";
 
-    public static final String S3_REGION = "manifeststore.s3.region";
+    public static final String S3_REGION = "fester.s3.region";
 
-    public static final String S3_BUCKET = "manifeststore.s3.bucket";
+    public static final String S3_BUCKET = "fester.s3.bucket";
 
     /**
      * Private constructor for the Constants class.
