+++ src/main/java/edu/ucla/library/iiif/manifeststore/Constants.java
@@ -1,5 +1,5 @@
 
-package edu.ucla.library.iiif.manifeststore;
+package edu.ucla.library.iiif.fester;
 
 /**
  * A class for package constants.
@@ -9,7 +9,7 @@ public final class Constants {
     /**
      * ResourceBundle file name for I18n messages.
      */
-    public static final String MESSAGES = "manifeststore_messages";
+    public static final String MESSAGES = "fester_messages";
 
     /**
      * The IP for an unspecified host.
