+++ src/main/java/edu/ucla/library/iiif/manifeststore/handlers/AbstractManifestHandler.java
@@ -1,13 +1,13 @@
 
-package edu.ucla.library.iiif.manifeststore.handlers;
+package edu.ucla.library.iiif.fester.handlers;
 
 import com.amazonaws.regions.RegionUtils;
 
 import info.freelibrary.util.Logger;
 import info.freelibrary.vertx.s3.S3Client;
 
-import edu.ucla.library.iiif.manifeststore.Config;
-import edu.ucla.library.iiif.manifeststore.MessageCodes;
+import edu.ucla.library.iiif.fester.Config;
+import edu.ucla.library.iiif.fester.MessageCodes;
 import io.vertx.core.Handler;
 import io.vertx.core.Vertx;
 import io.vertx.core.json.JsonObject;
