+++ src/main/webroot/success.html
@@ -10,8 +10,8 @@
 <meta name="viewport" content="width=device-width, initial-scale=1, shrink-to-fit=no">
 
 <link rel="stylesheet" href="/css/bootstrap.min.css">
-<link rel="stylesheet" href="/css/manifest-store-custom.css">
-<link rel="stylesheet" href="/css/manifest-store-response.css">
+<link rel="stylesheet" href="/css/fester-custom.css">
+<link rel="stylesheet" href="/css/fester-response.css">
 
 </head>
 
@@ -25,7 +25,7 @@
     <a class="navbar-brand" href="#">Fester</a>
     <div class="collapse navbar-collapse" id="navbarNavAltMarkup">
       <div class="navbar-nav">
-        <a class="nav-item nav-link active" href="/docs/manifest-store/">API Documentation</a>
+        <a class="nav-item nav-link active" href="/docs/fester/">API Documentation</a>
       </div>
     </div>
   </nav>
