+++ src/test/resources/test-config.properties
@@ -4,7 +4,7 @@
 #
 
 # S3 configuration options
-manifeststore.s3.bucket=${manifeststore.s3.bucket}
-manifeststore.s3.access_key=${manifeststore.s3.access_key}
-manifeststore.s3.secret_key=${manifeststore.s3.secret_key}
-manifeststore.s3.region=${manifeststore.s3.region}
+fester.s3.bucket=${fester.s3.bucket}
+fester.s3.access_key=${fester.s3.access_key}
+fester.s3.secret_key=${fester.s3.secret_key}
+fester.s3.region=${fester.s3.region}
