+++ src/test/scripts/pre-it-startup.sh
@@ -2,5 +2,5 @@
 
 # We have to provide a little wrapper script for exec-maven-plugin to background our job
 $* > /dev/null 2>&1 &
-echo $! > manifeststore-it.pid
+echo $! > fester-it.pid
 exit 0
